sql/modules/install.sql: make module paths relative to LedgerSMB root directory.
For reuse by the test script install-mycompany and future installers, prefix the
lines of sql/modules/install.sql with the path to those modules: sql/modules.

The script can now be run without changing to the modules directory first. Of
course, the same script can't be run in the modules directory now, either.

git-svn-id: svn://svn.code.sf.net/p/ledger-smb/code/trunk@2673 4979c152-3d1c-0410-bac9-87ea11338e46