Test case correction:
Behavior of from_to changed for 1.3.18.  The old behavior was to try to return a date in the user's datestyle, however since this is always used to generate dates for SQL queries, this was removed.  The dates are then expected to be run through ->format_date() or ->date() separately.

This was necessary to fix bugs in from_to handling with certain date formats. This corrects the test case (only) to reflect this change in usage.

git-svn-id: svn://svn.code.sf.net/p/ledger-smb/code/branches/1.3@4832 4979c152-3d1c-0410-bac9-87ea11338e46