INSTALL: clarify instructions on setting db password and setup.pl
The INSTALL instructions do not make clear that a password must be set
for the database superuser to use the setup.pl wizard.

This tripped me up on a clean Debian install and has puzzled others:
http://forums.ledgersmb.org/perl/lsmb/topic_show.pl?tid=28043

This commit clarifies the need for a password, suggests how to set one
and provides more explanation of how to use the setup.pl wizard to create
a new company.