Fix SF bug 1338 - Default account not being shown
This turned out to be a number of problems with System->Defaults form.

1) Default value was not being set for any of the account dropdowns,
as a result, simply opening the form and clicking save, could make silent
changes to the defaults without the user actively altering anything.

2) FX gain/loss dropdowns were being populated by searching for accounts
with an account_link description of 'FX_loss' or 'FX_gain'. However such
account_link descriptions do not exist, nor is there any way to set such
via the UI. This patch reverts to the 1.3 behaviour of listing all accounts
in the FX gain/loss account dropdowns.