Fix SF bug 1339 - Missing file error after saving bank reconciliation
Prior to this patch, code was pointing to a non-existent template file,
and doing this regardless of whether the requested save was valid.

If the save is not valid, we should display the reconciliation report
once again, as the error message suggests 'please update again'.

If the save is valid, we should return to the "Search Reconciliation
Reports" form. This is already in place and working as a separate
function in the module, so we can simply call that, rather than
coding a separate template.