Fix html errors in /UI/payments/payments_detail.html
* Remove label tag from around filter_to caption.
    This was causing incorrect display:
    was: "Filtering From: To: 2015-02-13 2015-02-15"
    now: "Filtering From: 2015-02-13 To: 2015-02-15"

* reformat for consistent indentation (previously a mix of spaces and tabs)

* add missing closing div tag to "date_filter_row"

* remove spurious closing td tag from "account_info" span

* change details_select span to a div (because it contains divs.
    A span cannot contain divs, only other inline tags.)

* invoice_list td element : fix misplaced closing td tag
    (content was outside closing tag)

* total_due_list td element : fix misplaced closing td tag
    (content was outside closing tag)

* invoice_detail_list td element : add missing closing tag

* container div : add missing closing tag

There are a number of places where a label tag is being used to add a caption
to a span. This is invalid html as label elements are only applicable to form
elements. I have left this as-is because changing it would affect the page
styling and in practice it is working OK.
(grafted from f8daa95758050570ecdd1338791c656cb5ae3121)