payment.pm - whitespace cleanup
I'd normally avoid patches that simply alter whitespace within a source file,
however payment.pm had sections without any indentation and a mixture of tab
and space indentation which made reading the code difficult.

Having taken the time to clean up the whitespace to aid my own understanding of
the code, I feel I might as well share the benefits of that with others.

I hope the clean-up makes it easier for others to follow the code and less likely
that new bugs will be introduced when making patches.

As ever, feedback and criticism welcome.