Fix Enable Submit for Bank Reconciliation in 1.5
My PR #2420 to fix bug #2419 was overly simplistic. It fixed one problem,
but broke something else by changing the sign in the calculation used to
decide whether the reconciliation report could be submitted.

In fact, the sign should be dependent on the account_category being reconciled.

The point of this code is to prevent submission of a reconciliation report if
the balance does not match. We are already calculating 'out_of_balance',
displayed as 'Variance', so there is no need to make a separate calculation.

This patch is designed as a minimal change to use the Variance to
determine whether the reconciliation report may be submitted.

This patch also ensures that the 'Variance' is compared to the number of
decimal places for money specified in lsmb's settings, whereas it was
previously hard-coded to compare pennies.