* Fix #2648: assembly stocking not in inventory activity report
This commit adds two columns to the inventory activity report: 'used' and
'assembled' (and renames a few others).
It also introduces a new test to verify that stocked assemblies and the
used parts are reported correctly in the inventory activity report.
As the numbers for the TS1 part (used for the assembly) were inserted
with the wrong sign -- this is corrected too.