* Remove checks for missing GATEWAY_INTERFACE variable
These checks were used to determine if the scripts were called in a
CLI (command line interface) scenario or not.  The solution to supporting
a CLI is not by littering the code with GATEWAY_INTERFACE checks, but
instead to create a true library which doesn't assume much about the
context it's running in, receiving the required inputs from its caller.