* Factor out codebase wide assumption of having creds in the %ENV hash
This commit adds an _auth key to the LedgerSMB and Form objects which
holds a LedgerSMB::Auth::DB object. To that extent, the module has been
converted from being an add-on for the LedgerSMB::Auth package to a
full-blown Moose class.
Since old-code still runs in a simulated CGI environment, that code still
can query the %ENV hash mid-execution. For clarity, this commit does not
depend on that possibility. Instead, the _auth key is initialized at
object construction time, just as with the PSGI code.