Add general dependency on Module::Runtime to cpanfile
A number of our modules require the Module::Runtime module, but it
was listed as a dependency only if the edi feature is enabled. This
change adds it as a general dependency.

In 1.6, the general dependency on Module::Runtime has been factored-out,
but I considered that change as too invasive for the stable 1.5 branch.
Better just to correct the dependency and leave code untouched.