* Fix parts selector ('dropdown') broken due to refactoring
Note that even after this change, the JSON is incorrectly
generated, IMO, but before this change there's reference
to a non-existing function. Without escaping, the JSON
encoder complains about 'covert_blessed' not being
enabled. (However, it seems to be, judging by the first
lines of code of the file...) More research necessary.