* Allow only part of a LOADORDER of changes to be applied
This commit is part of a bigger plan to migrate LedgerSMB 1.2/1.3 data and
SQL Ledger data to a fixed schema version. Bringing the data forward to the
lastest schema then is handled by LedgerSMB's own schema migrations.

Note that this plan increases chances for people to be able to migrate their
data as the migration scripts will no longer need to be continuously
adapted to the changes in the target schema (instead, we can start to
focus on changes in the source schema, as far as SQL Ledger is concerned).