* Collect HTML file validation failures into a single test result
Before, the number of reported HTML tests would depend both on the number
of tested files as well as on the number of reported failures. With this
change, the number of tests is exactly the same whether successes or failures.
Additionally, the criteria for non-reported validation errors are now
much more readable.