* Fix #2972: Forms sometimes seem to hang
The cause of the seeming hang is that the new pages are loaded when the
'hash' section of the URL changes. However, because any intermediate POST
requests don't change the hash, it can happen that two subsequent GET
form posts end up requesting the same URL. As a result, the second GET
won't be executed (as the 'hash' part of the window.location doesn't change).