* Fix reports having a hidden 'action' field
Our buttons post an 'action' field and having a hidden by the same
name produces an array of action values, which results in failure
to identify the correct entry-point.

Since the 'report' variable points to a DBObject instance, all
values from the request are imported into the report, meaning
the 'action' field from the request does too. We don't want
that.