Allow custom base url for Is_LSMB_running.sh test
Previously Is_LSMB_running.sh used a hard-coded url for ledgersmb,
meaning the Is_LSMB_running.sh test failed when a different port or url
was used, such as in the Docker ledgersmb-dev-lsmb container environment.

This change keeps the default localhost:5001 location, but allows a
different base url to be specified via the LSMB_BASE_URL environment
variable, as used in the standard lsmb Docker containers.