Document all authentication scenarios for AuthenticateSession and refactor
This commit moves the full responsibility for session management to
AuthenticateSession, including session instantianion (but session deletion
remains an open point to be addressed in a follow-up commit).
In order for AuthenticateSession to offer session creation (and later
destruction) services, callbacks are implemented.

AuthenticateSession *also* gets the responsibility to instantiate database
connections, because the session data determine which database to connect
to (in case there's a session).

Furthermore declare that login.pl::authenticate() and all setup.pl entry
points want a database connection (but not a session). login.pl's
authenticate() entrypoint wants to establish a session on top of the
database connection.

Delete session creation code from LedgerSMB::Session due to moving
the responsibility for sessions to AuthenticateSession.