Remove useless dblog file tests from xt/40-dbsetup.t
xt/40-dbsetup.t attempts to check and clean up various dblog files
from a temporary directory, but these files are never created, hence
the checks and cleanup are useless and are removed by this patch.

This patch is in preparation to enable InputOutput::RequireCheckedOpen
Perl::Critic policy on the tests.