Explicitly list Perl::Critic policies to apply
Previously we specified a combination of severity, theme, inclusion
and exclusion which meant different policies would be applied according
to what perl modules were installed on the system - unpredictable.

This patch explicitly lists the policies we wish to apply, raises an
error if they are not available on the system and separates this
configuration from code.