Constrain default password_duration value in database
Value of default password_duration is expected to be a positive
number, NULL or ''. Anything else will cause a fatal error when
trying to change a password.

The database uses a key/value table to store defaults, with the
value stored as a text field. This patch adds a check constraint
when the key is 'password_duration'.

During an upgrade, bad password_durations are dropped, effectively
replacing the bad value with the default 'infinite' password
validity. As the bad values don't work, no valid data is lost
by this process.