* Solve the running balance being incorrectly calculated
Note how PostgreSQL documents running window functions over the
virtual table being the result of WHERE, GROUP BY and HAVING processing
and note that that list excludes ORDER BY processing.  In order
to make sure we run the balance function over the same order
as is the result of the final query, we must include the same
ORDER BY condition in the window function.

Also note that a window function runs over the _current row *and* its
equivalents as considered under ORDER BY_ which is why we include
'ac.entry_id' in the order function: it's guaranteed to be unique.