Don't strip filename path in LedgerSMB::Mailer->attach method
Prompted by @hasorli's suggestion to replace a regex substitution
with File::Basename, I dug a little further...

We don't currently depend on path-stripping the filename argument
to this method. And conceptually, if the caller has explicitly
provided a filename, it is not our place to mangle it.

I have therfore remove the path stripping - we now use the provided
filename argument directly.