Fix template_format configuration options broken logic
* Fix broken logic for template_format config options
* Document template_format config options in LedgerSMB::Sysconfig
* Add tests for template_format config options

This fixes broken logic in processing the template_format
configuration options `template_latex`, `template_xls`,
`template_xlsx` and `template_ods`.

Output format options were always being displayed, regardless
of dependencies, when the respective configuration
option was set to "disabled".

In that instance, the dependency logic was not executed and the
raw "disabled" value was used as the configuration setting instead
of a processed true/false value. As "disabled" evaluates to true,
the format options were incorrectly displayed.