Conditionally enable/disable e-mail backup UI elements
If the `backup_email_from` configuration element is not defined,
attempting to e-mail a backup will fail with a user-unfriendly
error message.

This PR disables the e-mail backup UI elements unless the
`backup_email_from` configuration is set.

The facility to return backups to the browser remains enabled under
all circumstances and is made the default selection.