Add stdout/stderr capture after database run_file error
By design, PGObject::Util::DBAdmin croaks on error, which
prevents access to the captured log output. This commit adds a
run_file_with_logs() method, which wraps the run_file() method in
an eval to trap the error, then slurps and returns the log files.

This new method is applied inside the load_modules() method to
display a meaningful error message on failure.