Add dojo form validation to Password Duration setting
We have a database constraint for Password Duration. This commit
adds dojo form validation for this field so that a user error
is displayed if an invalid value is entered.

Previously a 'server error' box would be displayed on form
submission, which gave no clue as to what caused the issue.