Clarify user message after changing password via Preferences
Final part of fix for #3231 on master - corrects misleading error
message. Database constraint added by earlier commits prevent the
root cause.

Currently, when a user changes their password, any error is reported as
`Company does not exist`. This makes no sense, as we are logged in to
a company, so there must be another reason for the failure.

Message changed to more generic - and accurate - `Error changing password`.