Don't escape hash keys for templates
This fixed a regression between 1.5 and 1.6. Following refactoring
of the LedgerSMB::Template modules, both hash keys and values were
being escaped to create the `cleanvars` passed to the template.

This broke the PDF templates, causing their data rows to be empty.

I'd argue that applying a blanket escape to everything before the
template is flawed. We hae no idea whether the template is going to
use the data for user display, or internal processing. We should
therefore leave escaping up to the template. As things stand, we
have individual templates trying to un-escape data to get at the
original raw values. That's for another time...