Fix LedgerSMB::File initialisation of file_path property
The attribute was documented to be a Maybe[Str], but the initialization
function returns a File::Temp::Dir.

Created a separate private property to hold the File::Temp::Dir,
with initialisation to ensure the public read-only property `file_path`
contains the temporary directory path as a string.

Also Pod fixes to document return value of get_for_template() method.