Move LedgerSMB::File tests to separate part of test sequence
Tests were failing because the LedgerSMB::File tests are required to
make a temporary working copy of the test database. They were being
run in parallel with other tests which had opened the database,
which was preventing the copy being made.

They now have their own number prefix and are run sequentially.