* Fix #3676: Regression UTF-8 characters garbled
Note that all data submitted to the PSGI server is UTF-8 encoded
because the HTML5 standard requires data to be submitted in the
same encoding as the encoding of the form-containing document.

PSGI however, doesn't interpret encoding and simply interprets the
bytes in the input stream. Before, our code would instruct the CGI
parser to interpret input as UTF-8, which due to switching to PSGI
has been removed.

Add back the interpretation of the submitted values as UTF-8.