* Fix #3588: change 1.5/template_menu fails to apply
The problem we're fixing is that schema change 1.5/template_menu.sql
and 1.5/template_menu3.sql (by consequence) fail to apply.

The new script is much more careful to free positions in the menu and
check for existence of nodes than the previous scripts.

Note that the only way to add ACLs to menu items is through Roles.sql
because upon creation of a new database, the function lsmb__role()
hasn't been loaded into the schema yet, but it's return value in an
existing schema may differ from the standard 'lsmb_<database>__'.
Roles.sql takes care of all that for us.