* Refactor 'use' statements to comply with Critic's 'RequireExplicitInclusion' policy
Note that this commit applies to new code only and that there are
no plans to apply the policy to old code. In places where old code
'bleeds into' new code (dispatch points), it's already visible in
this commit that it's likely hard to make the policy work for old
code.