* Harmonize logging across tests: disable logging
Note that most tests used the logger as configured in
LedgerSMB::Sysconfig; however, that log config contains
the custom '%Z' token which is defined in one of the
Middlewares (which the tests don't use). As a result,
the tests generate warnings in some configurations.
This commit initializes logging to be off, thereby
eliminating the warnings regarding the undefined
logger token (as the config from LedgerSMB::Syslog
is no longer being used as initialization).