* Move 'root_doc' rendering to login.pl (from menu.pl)
Note that the 'root doc' really originally was the frameset and today
is just the skeleton that serves as the Single Page Application, with
the Dojo AMD hints for loading further assets.
To start loading the assets ASAP, we want to get the initial response
out as fast as we can. This one takes it from 350ms and over to less
than 55ms.