* Improve response times from the menu JSON request (350+ms -> ~100ms)
Note that the mentioned times are on my machine; however
the general direction seems to be significant.

Note also that the removed code was actually a no-op in
terms of 'encoding' the result, because the escape function
passed in did nothing but return the argument unchanged.

However, the consequence of the operation _was_ to 'unpack'
object references into plain hash references.