Remove PGOld duplicated base parameter processing
`LedgerSMB::PGOld` subclasses `PGObject::Simple`.

`LedgerSMB::PGOld->new()` accepts a `base` hashref parameter which
is used to initialise the new object's properties.

It was copying this base hashref to a new hash, before passing
to `PGObject::Simple`, which repeated the process, copying again
to a new hash to initialise the object.

This patch removes the redundant hash copy, passing the base hashref
directly to `PGObject::Simple`'s initialiser.

A unit test is added to ensure the object property initialisation is
performed correctly.