Fix LedgerSMB::oldcode::dispatch not passing arguments
This fixes a bug in `LedgerSMB::oldcode::dispatch` which is used to run entrypoints
into 'old' lsmb script modules.

The `Try::Tiny` documentation states that `@_` is unavailable inside a `try` block,
but we were trying to use `@_` inside such a block to pass arguments to the
entrypoint function.

`@_` is now copied at the start the `dispatch` function so it can be passed on
from within the `try` block.