Don't disconnect dbh when child process exits
`LedgerSMB::old_code::dispatch()` forks a child process to execute
'old code', before reading its output and translating it into a
standard PSGI response triplet.

By default, a `DBI` database handle is destroyed as part of that
module's cleanup when a process terminates.

In this case, the child process inherits and uses the same dbh
database handle as the parent process. When the child process exits,
a disconnect would normally be executed on that handle - destroying
the database handle still in use by the parent process.

This patch applies the `AutoInactiveDestroy` option to the database
handle, preventing a child process from destroying the parent's dbh.

See DBI documentation for more details on this option.