Don't explicitly disconnect dbh in App_State::cleanup()
We were explicitly calling `disconnect()` on the database handle as
part of the `LedgerSMB::App_State::cleanup()` function. If running
in a forked process, using the parent's database handle, this causes
the parent's database handle to be destroyed.

The `disconnect()` call is unnecessary and, if running in a forked
process, can prevent the parent process carrying on database operations.

Later in the `cleanup()` function, the database handle is set to `undef`.
This automatically disconnects the database handle under most situations,
but respects `DBI`'s `AutoInactiveDestroy` option to stop child processes
from closing a parent's database handle.