Remove LedgerSMB::PGOld->set_dbh() method
This patch removes a piece of old code and reduces dependency on the
global LedgerSMB::App_State, which we are trying to deprecate.

`LedgerSMB::PGOld` was overriding `PGObject::Simple`'s `set_dbh()`
method with one that used `LedgerSMB::App_State::DBH()` to initialise
the class's dbh.

This effect was not documented and removed any public way of explicitly
setting a different dbh.

The override has been removed. `LedgerSMB::PGOld->set_dbh()` now calls
the underlying `PGObject::Simple->set_dbh()` method, requiring a dbh
to be specified as an argument.

The overridden method was used in only two places within lsmb code.
The calls to it have been updated to use the dbh already present in
form/request objects, rather than using App_State.