Fix display_account_screen() to work without PGOld->set_dbh
This updates `LedgerSMB::Scripts::account->_display_account_screen()`
to correctly extract a database handle from both the types of input
arguments that it handles.

Previously `LedgerSMB::PGOld->set_dbh()` set the database handle
to use `LedgerSMB::App_State::DBH`, which we are trying to deprecate.

Now the database handle must be set either by a `dbh` argument
passed to the contructor (used when the base argument is a LedgerSMB
object) or as an explicit parameter to the `set_dbh` method.