Fix setup to work with updated PGOld
`LedgerSMB::PGOld` no longer uses `LedgerSMB::App_State::DBH()` to
initialise its database handle, which must now be passed-in as a
constructor parameter, or explicitly set.

Although setup appeared to be passing in a constructor containing
the `dbh` parameter, it was not being passed correctly. Objects
based on PGOld require a `base` key/value pair.

This is useless as the `$request` argument is completely ignored:

    `$user = LedgerSMB::DBObject::User->new($request);`

Instead, the database handle is now set explicitly using the
`set_dbh()` method.