Simplify LedgerSMB::PGOld, reduce App_State::DBH dependency (#3790)
* Drop redundant _DBH parameter from PGOld->new

Although undocumented, LedgerSMB::PGOld contains code which modifies
the provided arguments if a parameter `_DBH` is set. This is an
action-at-distance code pattern to surprise the unwary and unecessary
in current lsmb code. It is therefore removed.

* Remove useless parameter from LedgerSMB::PGOld->set_dbh call

`LedgerSMB::PGOld->set_dbh` does not accept arguments. It always uses
`LedgerSMB::App_State::DBH()` for initialisation.

In `LedgerSMB::DBObject::TransTemplate`, `LedgerSMB::PGOld->set_dbh`
was being called with a useless initialisation parameter. This has
been removed.

* Fix POD typos in LedgerSMB::DBObject::Account

* Use published interface to set PGObject::Simple db handle

LedgerSMB::PGOld->set_dbh() was manipulating an undocumented private
`_DBH` property of the `PGObject::Simple` class it inherits from.

We should not be manipulating the internal state of other classes.
Changed to use the documented setter method.

* Remove unnecessary subclassed method LedgerSMB::PGOld->dbh

`LedgerSMB::PGOld` subclassed the `PGObject::Simple` to return
the value of `LedgerSMB::App_State::DBH()' when called as a
simple function.

It was never called in this way, so this method has been
removed - calls can be handled by the base `PGObject::Simple`
method.

* Remove LedgerSMB::PGOld->set_dbh() method

This patch removes a piece of old code and reduces dependency on the
global LedgerSMB::App_State, which we are trying to deprecate.

`LedgerSMB::PGOld` was overriding `PGObject::Simple`'s `set_dbh()`
method with one that used `LedgerSMB::App_State::DBH()` to initialise
the class's dbh.

This effect was not documented and removed any public way of explicitly
setting a different dbh.

The override has been removed. `LedgerSMB::PGOld->set_dbh()` now calls
the underlying `PGObject::Simple->set_dbh()` method, requiring a dbh
to be specified as an argument.

The overridden method was used in only two places within lsmb code.
The calls to it have been updated to use the dbh already present in
form/request objects, rather than using App_State.

* Remove useless test of LedgerSMB::Setting->dbh()

`LedgerSMB::Setting->dbh()` is never called in lsmb code, so the
test for that is removed.

A test is added to ensure that `LedgerSMB::Setting->get` is using
the expected database handle. It should explicitly and directly use
`LedgerSMB::App_State::DBH`.

* Fix display_account_screen() to work without PGOld->set_dbh

This updates `LedgerSMB::Scripts::account->_display_account_screen()`
to correctly extract a database handle from both the types of input
arguments that it handles.

Previously `LedgerSMB::PGOld->set_dbh()` set the database handle
to use `LedgerSMB::App_State::DBH`, which we are trying to deprecate.

Now the database handle must be set either by a `dbh` argument
passed to the contructor (used when the base argument is a LedgerSMB
object) or as an explicit parameter to the `set_dbh` method.

* Fix pod typo in LedgerSMB::DBObject::Account

No code changes.

* Fix setup to work with updated PGOld

`LedgerSMB::PGOld` no longer uses `LedgerSMB::App_State::DBH()` to
initialise its database handle, which must now be passed-in as a
constructor parameter, or explicitly set.

Although setup appeared to be passing in a constructor containing
the `dbh` parameter, it was not being passed correctly. Objects
based on PGOld require a `base` key/value pair.

This is useless as the `$request` argument is completely ignored:

    `$user = LedgerSMB::DBObject::User->new($request);`

Instead, the database handle is now set explicitly using the
`set_dbh()` method.

* Fix BDD tests to work with new LedgerSMB::PGOld

`LedgerSMB::PGOld` no longer relies on `LedgerSMB::App_State::DBH()`
for its database handle, which must now be set explicitly.

* Fix BDD tests to work with new LedgerSMB::PGOld - part 2

`LedgerSMB::PGOld` no longer relies on `LedgerSMB::App_State::DBH()`
for its database handle, which must now be set explicitly.

Fixing a further instance for parts tests.

* Fix BDD tests to work with new LedgerSMB::PGOld - part 2

`LedgerSMB::PGOld` no longer relies on `LedgerSMB::App_State::DBH()`
for its database handle, which must now be set explicitly.

Fixing error in previous commit.