Use Plack::Request to initialise LedgerSMB object
Rather thank unpacking selected parts of a `Plack::Request` object to initialise
a new `LedgerSMB` object, pass the whole unmodified `Plack::Request` object.

This is then a standard, documented interface.

This opens the way to simplifying the `LedgerSMB`, making the request object
one of it's properties and providing attribute getters/setters so that users
of the `LedgerSMB` module no longer directly manipulate its internal state.