* Demonstrate a use-case for the inputfield mapper
Note that the structure of the fields in the payment module is highly
complex. In order to be able to use the inputfield mapper, this commit
simplifies the structure (naming) of the input tags a bit in order to
allow more straightforward mapping.

Additionally, the new code that uses the mapper is simplified, but the
simplification isn't restricted to just the 'pre_bulk_post_report'
function: other functions parse the values from the bulk post interface
too; those can be addressed using the same mapper (as long as the code
is in the payment.pl script). Since the mapper can be re-used, code in
general (in this module) will become more clear and more concise.