Add Timecard - make date required field
When adding a timecard, the first step is to complete the
"Timecard Criteria" form. If the `Starting Date` field is left
blank, an error relating to operations on an undefined value
is displayed.

This patch marks the `Starting Date` field as `required` to
provide user feedback if the field is not filled-in and to avoid
triggering the error.