Fix LedgerSMB::Batch->create returns false, update pod
`LedgerSMB::Batch->create` is intended to return the `id` of the
newly created batch.

The code was always returning an `undef` value as it was attempting
to return a non-existent hash key.

It now correctly returns the `id` of the newly created batch.

Pod updated to better document get() and create() methods.