* Implement a mechanism for fixes to schema changes
Note that there are still strict conditions for this to work
e.g. that the original schema changes still need to be stored
together with the new schema change files.
However, this new situation is a lot less rigid than the one
we had before where schema change files had to be completely
immutable.