* Aggregate transaction totals before combining with checkpoints
This change eliminates the pattern of 'exploding' the
aggregates from account_checkpoint to all transaction
rows of acc_trans, requiring the aggregates to be
aggregated by an identity function.

Instead, the pattern now aggregates the transactions
before joining with the checkpoints. The result is
that two tables, both aggregated on 'account_id',
need to be joined without aggregation afterwards.