Fix create new Reverse Payments batch
Previous changes had broken the ability to create a new
Reverse Payments batch and add vouchers to it. Fixed by
this PR.

This PR changes handling of the `$request` object. Rather than
passing the whole request through two instances of `LedgerSMB::Batch`,
letting it modify the request, then deleting an unwanted key, we
preserve the original request object. This ensures that the downstream
modules receive the full `LedgerSMB` object reference and can call
its methods.

The call to `LedgerSMB::Batch->create` to create a new batch is now
handled separately, explicity passing arguments in from
the request and explicitly writing just the wanted output back to
the request. This avoids copying the whole request multiple times
and better documents the flow of arguments.