* Factor UI template rendering into a separate module
This change is a step in the direction of completely separating document and UI
template processing into their own modules, thereby reducing complexity in both.

The UI template is a singleton pattern, setting up the renderer once (per worker)
and re-using the renderer thereafter (until the worker gets torn down).

This pattern allows faster response generation for those pages/entrypoints which
use it.