Remove unused fields from Filter Payments screen
The Filter Payments screen is used when adding payment or receipt
vouchers to a batch. It provides fields to filter the
payments/recipts before presenting them for selection.

This PR removes a number of fields from this screen which were
unused because there is no backend functionality to support them:

  * categories
  * department
  * businesses
  * date_from
  * date_to

The underlying `payment_get_all_contact_invoices` database procedure
appears to support filtering on these fields, but on close inspection,
the query it runs does not take notice of these parameters.

It is unclear in any case exactly what date the from/to filters should
operate on - due date, tax point, entry date - it's ambiguous.

Removing these fields presents only the functionality we have in place,
and are able to test, without the distraction of features
yet to be implemented. They can easily be added back as and when anybody
requires these features and is willing to implement them.