Fix LedgerSMB::DBObject::Payment->get_default_currency()
While `LedgerSMB::DBObject::Payment->get_default_currency()` was
returning the expected three-character currency code, it was not
setting the object's `default_currency` property to this value,
which is the expected behaviour.

This PR adds documentation to clarify the method's function and
return value. It adds unit tests to confirm its operation.