Batch_Overview
Property `class_id` can be `undef` to indicate that batches of
any class should be returned.

The Moose property was defined to allow only `Int` values, so
this property could not be explicitly set to `undef`.

Moose property now altered to `isa => 'Maybe[Int]'` so that it
can be explicitly set to `undef`.