SQL fix batch__search - interpretation of in_approved parameter
SQL function `batch__search` is documented that setting any of the
filtering input parameters to `NULL` will accept any value for the
respective field.

For the `in_approved` input parameter, the query was interpreting
both `NULL` and `FALSE` to mean only unapproved batches should be
returned.

The query has been fixed to match the documentation:

* in_approved = NULL  -- both approved and unapproved batches
* in_approved = TRUE  -- only approved batches
* in_approved = FALSE -- only unapproved batches