Fix Transaction Approval-Batches description search
The primary purpose of this PR is to fix a broken search for batches
based on their `Description`.

The search filter screen was being shared with a search filter for Drafts,
which uses different terminology and underlying parameter names, referring
to `Reference` rather than `Description`.

Rather than add yet more conditional blocks to a shared template, this PR
separates the Batches and Drafts filter screens into different and dedicated
html templates, without need for conditionals and their associated control
parameters.

Additionally, the `module_name` parameter is removed from the menu url used
to display these reports. This is unnecessary and was causing an additional
code block to be pointlessly executed.

Further simplification is possible in future. Both filter reports are
delivered through the `start_report` routine, which calls 11 database
functions in turn, all of which are unnecessary for these screens.