Add sql function `batch_get_class_name`
The database has a function `batch_get_class_id` to obtain the batch
class id corresponding to a particular batch class name.

This PR adds the inverse function, to return the batch class name for
a specified batch class id.

This is needed to enable the Batch Search report to display the name
of the batch class being filtered in the heading section.