Standardise Search Reconciliation Reports UI
`Search Reconciliation Reports` is part of the `Transaction Approval`
menu hierarchy. This PR applies html structure and classes to match
the other members of this category.

Title is now part of form and displays with the same styling as
other titles (by default White text on black background).

Field labels displayed without trailing colon.

Input fields and labels are aligned.