Extend Payments Batch BDD tests
Extends BDD testing of payments vouchers. Now completes the process of
adding a payment into a voucher and checks that when listing batches
for approval, the total amount is correct.

This exercises the functions `pre_bulk_post_report` and `post_bulk`
which are to be modified by PR #3795 and therefore provides testing
for that PR.