Remove error-causing warning from Reconciliation Report generation
During display of reconciliation reports, a warning was logged if
the previous reconciliation report balance did not exactly match
the calculated books balance at that point.

This warning has been removed.

It caused a fatal error on some versions of perl if there was
no previous reconciliation report for the account being reconciled
through trying to compare an undefined value.

It did not cater for fractional-penny differences common in
accounts with foreign exchange transactions.

As this warning was only visible by viewing the error log and is
no indication that anyting is actually wrong, this seems to be
unnecessary.